+++ pollo-fresco/database/migrations/2026_03_04_000000_create_comprobantes_venta_tables.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('ventas', function (Blueprint $table) {
             $table->bigIncrements('comprobante_venta_id');
-            $table->unsignedBigInteger('usuario_id');
+            $table->unsignedInteger('usuario_id');
             $table->string('tipo_comprobante', 20);
             $table->string('serie', 10);
             $table->string('numero', 20);
